+++ src/lightning_app/frontend/panel/__init__.py
@@ -1,4 +1,4 @@
-"""The PanelFrontend and AppStateWatcher makes it easy to create lightning apps with the Panel data app
+"""The PanelFrontend and AppStateWatcher make it easy to create Lightning Apps with the Panel data app
 framework."""
 from lightning_app.frontend.panel.panel_frontend import PanelFrontend
 from lightning_app.frontend.utilities.app_state_watcher import AppStateWatcher