+++ src/lightning_app/frontend/utilities/app_state_watcher.py
@@ -1,7 +1,7 @@
 """The AppStateWatcher enables a Frontend to.
 
-- subscribe to app state changes
-- to access and change the app state.
+- subscribe to App state changes
+- to access and change the App state.
 
 This is particularly useful for the PanelFrontend but can be used by other Frontends too.
 """
@@ -23,8 +23,8 @@ _logger = logging.getLogger(__name__)
 class AppStateWatcher(param.Parameterized):
     """The AppStateWatcher enables a Frontend to:
 
-    - Subscribe to any app state changes.
-    - To access and change the app state from the UI.
+    - Subscribe to any App state changes.
+    - To access and change the App state from the UI.
 
     This is particularly useful for the PanelFrontend, but can be used by
     other Frontend's too.